# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Trip Builder API'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/login:
    post:
      summary: Authenticate
      operationId: authenticate
      description: 'Authenticate user and returns a `bearer` token.'
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: kwest@example.org
                password:
                  type: string
                  description: ''
                  example: 'thoEW:d(;p@?}a'
              required:
                - email
                - password
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []